Add delay-rate library code (#52)
* New docs

* Implement delay-rate

* Ignore

* Lint

* Restore type check

* style: pre-commit fixes

* Lint

---------
