+++ .gitignore
@@ -160,4 +160,5 @@ jolly_roger/_version.py
 subtractor_uvw.ipynb
 subtractor.ipynb
 scratch_plots.ipynb
+rate_scratch.ipynb
 .vscode/settings.json
